+++ mne/label.py
@@ -10,14 +10,17 @@ from os import path as op
 import os
 import copy as cp
 import re
+from warnings import warn
 
 import numpy as np
 from scipy import linalg, sparse
 
+from .fixes import in1d
 from .utils import (get_subjects_dir, _check_subject, logger, verbose,
                     deprecated)
 from .source_estimate import (_read_stc, mesh_edges, mesh_dist, morph_data,
                               SourceEstimate, spatial_src_connectivity)
+from .source_space import add_source_space_distances
 from .surface import read_surface, fast_cross_3d
 from .source_space import SourceSpaces
 from .parallel import parallel_func, check_n_jobs
@@ -188,18 +191,22 @@ class Label(object):
             raise ValueError('hemi must be a string, not %s' % type(hemi))
         vertices = np.asarray(vertices)
         if np.any(np.diff(vertices.astype(int)) <= 0):
-            raise ValueError('Vertices must be ordered in increasing '
-                             'order.')
+            raise ValueError('Vertices must be ordered in increasing order.')
+
         if color is not None:
             from matplotlib.colors import colorConverter
             color = colorConverter.to_rgba(color)
 
         if values is None:
             values = np.ones(len(vertices))
+        else:
+            values = np.asarray(values)
+
         if pos is None:
             pos = np.zeros((len(vertices), 3))
-        values = np.asarray(values)
-        pos = np.asarray(pos)
+        else:
+            pos = np.asarray(pos)
+
         if not (len(vertices) == len(values) == len(pos)):
             err = ("vertices, values and pos need to have same length (number "
                    "of vertices)")
@@ -347,6 +354,63 @@ class Label(object):
             The copied label.
         """
         return cp.deepcopy(self)
+
+    def fill(self, src, name=None):
+        """Fill the surface between sources for a label defined in source space
+
+        Parameters
+        ----------
+        src : SourceSpaces
+            Source space in which the label was defined. If a source space is
+            provided, the label is expanded to fill in surface vertices that
+            lie between the vertices included in the source space. For the
+            added vertices, ``pos`` is filled in with positions from the
+            source space, and ``values`` is filled in from the closest source
+            space vertex.
+        name : None | str
+            Name for the new Label (default is self.name).
+
+        Returns
+        -------
+        label : Label
+            The label covering the same vertices in source space but also
+            including intermediate surface vertices.
+        """
+        # find source space patch info
+        if self.hemi == 'lh':
+            hemi_src = src[0]
+        elif self.hemi == 'rh':
+            hemi_src = src[1]
+
+        if not np.all(in1d(self.vertices, hemi_src['vertno'])):
+            msg = "Source space does not contain all of the label's vertices"
+            raise ValueError(msg)
+
+        nearest = hemi_src['nearest']
+        if nearest is None:
+            msg = ("Computing patch info for source space, this can take "
+                   "a while. In order to avoid this in the future, run "
+                   "mne.add_source_space_distances() on the source space "
+                   "and save it.")
+            logger.warn(msg)
+            add_source_space_distances(src)
+            nearest = hemi_src['nearest']
+
+        # find new vertices
+        include = in1d(nearest, self.vertices, False)
+        vertices = np.nonzero(include)[0]
+
+        # values
+        nearest_in_label = np.digitize(nearest[vertices], self.vertices, True)
+        values = self.values[nearest_in_label]
+        # pos
+        pos = hemi_src['rr'][vertices]
+
+        if name is None:
+            name = self.name
+        label = Label(vertices, pos, values, self.hemi, self.comment, name,
+                      None, self.subject, self.color)
+        return label
 
     @verbose
     def smooth(self, subject=None, smooth=2, grade=None,
@@ -925,7 +989,7 @@ def label_sign_flip(label, src):
     return flip
 
 
-def stc_to_label(stc, src=None, smooth=5, connected=False, subjects_dir=None):
+def stc_to_label(stc, src=None, smooth=None, connected=False, subjects_dir=None):
     """Compute a label from the non-zero sources in an stc object.
 
     Parameters
@@ -936,13 +1000,18 @@ def stc_to_label(stc, src=None, smooth=5, connected=False, subjects_dir=None):
         The source space over which the source estimates are defined.
         If it's a string it should the subject name (e.g. fsaverage).
         Can be None if stc.subject is not None.
-    smooth : int
-        Number of smoothing steps to use.
+    smooth : bool
+        Fill in vertices on the cortical surface that are not in the source
+        space based on the closest source space vertex (requires
+        src to be a SourceSpace). The default is currently to smooth with a
+        deprecated method, and will change to True in v0.9 (i.e., the parameter
+        should be explicitly specified as boolean until then to avoid a
+        deprecation warning).
     connected : bool
         If True a list of connected labels will be returned in each
         hemisphere. The labels are ordered in decreasing order depending
         of the maximum value in the stc.
-    subjects_dir : string, or None
+    subjects_dir : str | None
         Path to SUBJECTS_DIR if it is not set in the environment.
 
     Returns
@@ -966,10 +1035,29 @@ def stc_to_label(stc, src=None, smooth=5, connected=False, subjects_dir=None):
     if not isinstance(stc, SourceEstimate):
         raise ValueError('SourceEstimate should be surface source estimates')
 
+    if not isinstance(smooth, bool):
+        if smooth is None:
+            msg = ("The smooth parameter was not explicitly specified. The "
+                   "default behavior of stc_to_label() will change in v0.9 "
+                   "to filling the label using source space patch "
+                   "information. In order to avoid this warning, set smooth "
+                   "to a boolean explicitly.")
+            smooth = 5
+        else:
+            msg = ("The smooth parameter of stc_to_label() was specified as "
+                   "int. This value is deprecated and will raise an error in "
+                   "v0.9. In order to avoid this warning, set smooth to a "
+                   "boolean.")
+        warn(msg, DeprecationWarning)
+
     if isinstance(src, string_types):
         if connected:
             raise ValueError('The option to return only connected labels is '
                              'only available if source spaces are provided.')
+        if isinstance(smooth, bool) and smooth:
+            msg = ("stc_to_label with smooth='patch' requires src to be an "
+                   "instance of SourceSpace")
+            raise ValueError(msg)
         subjects_dir = get_subjects_dir(subjects_dir)
         surf_path_from = op.join(subjects_dir, src, 'surf')
         rr_lh, tris_lh = read_surface(op.join(surf_path_from,
@@ -1035,18 +1123,20 @@ def stc_to_label(stc, src=None, smooth=5, connected=False, subjects_dir=None):
             colors = _n_colors(len(clusters))
             for c, color in zip(clusters, colors):
                 idx_use = c
-                for k in range(smooth):
-                    e_use = e[:, idx_use]
-                    data1 = e_use * np.ones(len(idx_use))
-                    idx_use = np.where(data1)[0]
+                if isinstance(smooth, bool) and smooth:
+                    label = Label(idx_use, this_rr[idx_use], None, hemi,
+                                  'Label from stc', subject=subject,
+                                  color=color).fill(src)
+                else:
+                    for k in range(smooth):
+                        e_use = e[:, idx_use]
+                        data1 = e_use * np.ones(len(idx_use))
+                        idx_use = np.where(data1)[0]
 
-                label = Label(vertices=idx_use,
-                              pos=this_rr[idx_use],
-                              values=np.ones(len(idx_use)),
-                              hemi=hemi,
-                              comment='Label from stc',
-                              subject=subject,
-                              color=color)
+                    label = Label(idx_use, this_rr[idx_use], None, hemi,
+                                  'Label from stc', subject=subject,
+                                  color=color)
+
                 this_labels.append(label)
 
             if not connected: